Version beta 2, incluye registro de adicionales y deducciones ademas la informacion d e6 proyectos ya registradas en la base de datos

+++ database/migrations/2018_03_06_132832_create_lpresupuesto_table.php
@@ -15,7 +15,7 @@ class CreateLpresupuestoTable extends Migration
     {
         Schema::create('gcolpresupuesto', function (Blueprint $table) {
             $table->increments('lprId');
-            $table->integer('lprOrderItem')->unsigned();
+            $table->decimal('lprOrderItem')->unsigned();
             $table->string('lprCodeItem',20)->nullable();
             $table->string('lprDescriptionItem');
             $table->boolean('lprIsProportion');
